+++ src/multi_vector_simulator/D1_model_components.py
@@ -15,7 +15,8 @@ Module D1 includes all functions that are required to build an oemof model with 
 
 import logging
 
-import oemof.solph as solph
+import pandas as pd
+from oemof import solph
 
 from multi_vector_simulator.utils.constants_json_strings import (
     VALUE,
@@ -274,6 +275,34 @@ def storage(model, dict_asset, **kwargs):
     - test_storage_fix()
 
     """
+
+    # Make sure the initial storage level is within the max and min values
+    initial_storage_level = dict_asset[STORAGE_CAPACITY][SOC_INITIAL][VALUE]
+    min_storage_level = dict_asset[STORAGE_CAPACITY][SOC_MIN][VALUE]
+    max_storage_level = dict_asset[STORAGE_CAPACITY][SOC_MAX][VALUE]
+
+    if initial_storage_level is not None:
+        if pd.isna(initial_storage_level):
+            dict_asset[STORAGE_CAPACITY][SOC_INITIAL][VALUE] = min_storage_level
+
+        if initial_storage_level < min_storage_level:
+            dict_asset[STORAGE_CAPACITY][SOC_INITIAL][VALUE] = min_storage_level
+            logging.warning(
+                f"The initial storage level of the battery asset {dict_asset[LABEL]} was below the minimal allowed value ({initial_storage_level} < {min_storage_level}), the initial level was ajusted to be equal to the minimum, please check your input files."
+            )
+        elif initial_storage_level > max_storage_level:
+            dict_asset[STORAGE_CAPACITY][SOC_INITIAL][VALUE] = max_storage_level
+            logging.warning(
+                f"The initial storage level of the battery asset {dict_asset[LABEL]} was above the maximal allowed value ({initial_storage_level} > {max_storage_level}), the initial level was ajusted to be equal to the maximum, please check your input files."
+            )
+    else:
+        if not isinstance(min_storage_level, float):
+            raise ValueError(
+                f"At the moment it is not possible to use multiple values of min_storage level"
+            )
+            min_storage_level = min_storage_level[0]
+        dict_asset[STORAGE_CAPACITY][SOC_INITIAL][VALUE] = min_storage_level
+
     check_optimize_cap(
         model,
         dict_asset,
@@ -326,6 +355,7 @@ def sink(model, dict_asset, **kwargs):
     """
     if TIMESERIES in dict_asset:
         sink_non_dispatchable(model, dict_asset, **kwargs)
+
     else:
         sink_dispatchable_optimize(model, dict_asset, **kwargs)
 
@@ -629,7 +659,7 @@ def transformer_constant_efficiency_fix(model, dict_asset, **kwargs):
         }
 
     if missing_dispatch_prices_or_efficiencies is None:
-        t = solph.Transformer(
+        t = solph.components.Converter(
             label=dict_asset[LABEL],
             inputs=inputs,
             outputs=outputs,
@@ -806,7 +836,7 @@ def transformer_constant_efficiency_optimize(model, dict_asset, **kwargs):
         }
 
     if missing_dispatch_prices_or_efficiencies is None:
-        t = solph.Transformer(
+        t = solph.components.Converter(
             label=dict_asset[LABEL],
             inputs=inputs,
             outputs=outputs,
@@ -991,15 +1021,16 @@ def source_non_dispatchable_fix(model, dict_asset, **kwargs):
     """
     outputs = {
         kwargs[OEMOF_BUSSES][dict_asset[OUTFLOW_DIRECTION]]: solph.Flow(
-            label=dict_asset[LABEL],
             fix=dict_asset[TIMESERIES],
             nominal_value=dict_asset[INSTALLED_CAP][VALUE],
             variable_costs=dict_asset[DISPATCH_PRICE][VALUE],
-            emission_factor=dict_asset[EMISSION_FACTOR][VALUE],
+            custom_attributes=dict(emission_factor=dict_asset[EMISSION_FACTOR][VALUE]),
         )
     }
 
-    source_non_dispatchable = solph.Source(label=dict_asset[LABEL], outputs=outputs)
+    source_non_dispatchable = solph.components.Source(
+        label=dict_asset[LABEL], outputs=outputs
+    )
 
     model.add(source_non_dispatchable)
     kwargs[OEMOF_SOURCE].update({dict_asset[LABEL]: source_non_dispatchable})
@@ -1034,7 +1065,6 @@ def source_non_dispatchable_optimize(model, dict_asset, **kwargs):
         existing = dict_asset[INSTALLED_CAP][VALUE]
     outputs = {
         kwargs[OEMOF_BUSSES][dict_asset[OUTFLOW_DIRECTION]]: solph.Flow(
-            label=dict_asset[LABEL],
             fix=dict_asset[TIMESERIES_NORMALIZED],
             investment=solph.Investment(
                 ep_costs=dict_asset[SIMULATION_ANNUITY][VALUE]
@@ -1046,11 +1076,13 @@ def source_non_dispatchable_optimize(model, dict_asset, **kwargs):
             variable_costs=dict_asset[DISPATCH_PRICE][VALUE]
             / dict_asset[TIMESERIES_PEAK][VALUE],
             # add emission_factor for emission contraint
-            emission_factor=dict_asset[EMISSION_FACTOR][VALUE],
+            custom_attributes=dict(emission_factor=dict_asset[EMISSION_FACTOR][VALUE]),
         )
     }
 
-    source_non_dispatchable = solph.Source(label=dict_asset[LABEL], outputs=outputs)
+    source_non_dispatchable = solph.components.Source(
+        label=dict_asset[LABEL], outputs=outputs
+    )
 
     model.add(source_non_dispatchable)
     kwargs[OEMOF_SOURCE].update({dict_asset[LABEL]: source_non_dispatchable})
@@ -1079,7 +1111,6 @@ def source_dispatchable_optimize(model, dict_asset, **kwargs):
     if TIMESERIES_NORMALIZED in dict_asset:
         outputs = {
             kwargs[OEMOF_BUSSES][dict_asset[OUTFLOW_DIRECTION]]: solph.Flow(
-                label=dict_asset[LABEL],
                 max=dict_asset[TIMESERIES_NORMALIZED],
                 investment=solph.Investment(
                     ep_costs=dict_asset[SIMULATION_ANNUITY][VALUE]
@@ -1091,10 +1122,14 @@ def source_dispatchable_optimize(model, dict_asset, **kwargs):
                 variable_costs=dict_asset[DISPATCH_PRICE][VALUE]
                 / dict_asset[TIMESERIES_PEAK][VALUE],
                 # add emission_factor for emission contraint
-                emission_factor=dict_asset[EMISSION_FACTOR][VALUE],
+                custom_attributes=dict(
+                    emission_factor=dict_asset[EMISSION_FACTOR][VALUE]
+                ),
             )
         }
-        source_dispatchable = solph.Source(label=dict_asset[LABEL], outputs=outputs,)
+        source_dispatchable = solph.components.Source(
+            label=dict_asset[LABEL], outputs=outputs,
+        )
     else:
         if TIMESERIES in dict_asset:
             logging.info(
@@ -1107,7 +1142,6 @@ def source_dispatchable_optimize(model, dict_asset, **kwargs):
             )
         outputs = {
             kwargs[OEMOF_BUSSES][dict_asset[OUTFLOW_DIRECTION]]: solph.Flow(
-                label=dict_asset[LABEL],
                 investment=solph.Investment(
                     ep_costs=dict_asset[SIMULATION_ANNUITY][VALUE],
                     existing=dict_asset[INSTALLED_CAP][VALUE],
@@ -1115,10 +1149,15 @@ def source_dispatchable_optimize(model, dict_asset, **kwargs):
                 ),
                 variable_costs=dict_asset[DISPATCH_PRICE][VALUE],
                 # add emission_factor for emission contraint
-                emission_factor=dict_asset[EMISSION_FACTOR][VALUE],
+                custom_attributes=dict(
+                    emission_factor=dict_asset[EMISSION_FACTOR][VALUE],
+                ),
             )
         }
-        source_dispatchable = solph.Source(label=dict_asset[LABEL], outputs=outputs,)
+        print(dict_asset[LABEL])
+        source_dispatchable = solph.components.Source(
+            label=dict_asset[LABEL], outputs=outputs
+        )
     model.add(source_dispatchable)
     kwargs[OEMOF_SOURCE].update({dict_asset[LABEL]: source_dispatchable})
     logging.debug(
@@ -1146,15 +1185,18 @@ def source_dispatchable_fix(model, dict_asset, **kwargs):
     if TIMESERIES_NORMALIZED in dict_asset:
         outputs = {
             kwargs[OEMOF_BUSSES][dict_asset[OUTFLOW_DIRECTION]]: solph.Flow(
-                label=dict_asset[LABEL],
                 max=dict_asset[TIMESERIES_NORMALIZED],
-                existing=dict_asset[INSTALLED_CAP][VALUE],
+                nominal_value=dict_asset[INSTALLED_CAP][VALUE],
                 variable_costs=dict_asset[DISPATCH_PRICE][VALUE],
                 # add emission_factor for emission contraint
-                emission_factor=dict_asset[EMISSION_FACTOR][VALUE],
+                custom_attributes=dict(
+                    emission_factor=dict_asset[EMISSION_FACTOR][VALUE]
+                ),
             )
         }
-        source_dispatchable = solph.Source(label=dict_asset[LABEL], outputs=outputs,)
+        source_dispatchable = solph.components.Source(
+            label=dict_asset[LABEL], outputs=outputs,
+        )
     else:
         if TIMESERIES in dict_asset:
             logging.info(
@@ -1167,12 +1209,13 @@ def source_dispatchable_fix(model, dict_asset, **kwargs):
             )
         outputs = {
             kwargs[OEMOF_BUSSES][dict_asset[OUTFLOW_DIRECTION]]: solph.Flow(
-                label=dict_asset[LABEL],
-                existing=dict_asset[INSTALLED_CAP][VALUE],
+                nominal_value=dict_asset[INSTALLED_CAP][VALUE],
                 variable_costs=dict_asset[DISPATCH_PRICE][VALUE],
             )
         }
-        source_dispatchable = solph.Source(label=dict_asset[LABEL], outputs=outputs,)
+        source_dispatchable = solph.components.Source(
+            label=dict_asset[LABEL], outputs=outputs,
+        )
     model.add(source_dispatchable)
     kwargs[OEMOF_SOURCE].update({dict_asset[LABEL]: source_dispatchable})
     logging.debug(
@@ -1206,7 +1249,6 @@ def sink_dispatchable_optimize(model, dict_asset, **kwargs):
         index = 0
         for bus in dict_asset[INFLOW_DIRECTION]:
             inputs[kwargs[OEMOF_BUSSES][bus]] = solph.Flow(
-                label=dict_asset[LABEL],
                 variable_costs=dict_asset[DISPATCH_PRICE][VALUE][index],
                 investment=solph.Investment(),
             )
@@ -1214,14 +1256,13 @@ def sink_dispatchable_optimize(model, dict_asset, **kwargs):
     else:
         inputs = {
             kwargs[OEMOF_BUSSES][dict_asset[INFLOW_DIRECTION]]: solph.Flow(
-                label=dict_asset[LABEL],
                 variable_costs=dict_asset[DISPATCH_PRICE][VALUE],
                 investment=solph.Investment(),
             )
         }
 
     # create and add excess electricity sink to micro_grid_system - variable
-    sink_dispatchable = solph.Sink(label=dict_asset[LABEL], inputs=inputs,)
+    sink_dispatchable = solph.components.Sink(label=dict_asset[LABEL], inputs=inputs,)
     model.add(sink_dispatchable)
     kwargs[OEMOF_SINK].update({dict_asset[LABEL]: sink_dispatchable})
     logging.debug(
@@ -1263,7 +1304,7 @@ def sink_non_dispatchable(model, dict_asset, **kwargs):
         }
 
     # create and add demand sink to micro_grid_system - fixed
-    sink_demand = solph.Sink(label=dict_asset[LABEL], inputs=inputs,)
+    sink_demand = solph.components.Sink(label=dict_asset[LABEL], inputs=inputs,)
     model.add(sink_demand)
     kwargs[OEMOF_SINK].update({dict_asset[LABEL]: sink_demand})
     logging.debug(