+++ tests/test_C2_economic_functions.py
@@ -160,6 +160,19 @@ def test_get_replacement_costs_one_reinvestment():
     assert replacement_costs == exp
 
 
+def test_get_replacement_costs_one_reinvestment_age_asset_equal_asset_lifetime():
+    replacement_costs = C2.get_replacement_costs(
+        age_of_asset=10,
+        project_lifetime=20,
+        asset_lifetime=10,
+        first_time_investment=550,
+        discount_factor=0.1,
+    )
+    # Investment in year 5 - present value of residual value = Investment in year 5 / Asset lifetime * used years
+    exp = 762.0488091862422
+    assert replacement_costs == exp
+
+
 def test_get_replacement_costs_no_reinvestment_residual():
     replacement_costs = C2.get_replacement_costs(
         age_of_asset=5,