+++ tests/test_E1_process_results.py
@@ -1,12 +1,12 @@
 import pandas as pd
-from pandas.util.testing import assert_series_equal
+from pandas.testing import assert_series_equal
 
 import os
 import numpy as np
 import logging
 import shutil
 import mock
-import oemof.solph as solph
+from oemof import solph
 import pickle
 
 import multi_vector_simulator.A0_initialization as A0