+++ tests/test_benchmark_scenarios.py
@@ -14,7 +14,7 @@ import pandas as pd
 import numpy as np
 import pytest
 from pytest import approx
-from pandas.util.testing import assert_series_equal
+from pandas.testing import assert_series_equal
 
 from multi_vector_simulator.cli import main
 from multi_vector_simulator.server import run_simulation
@@ -152,10 +152,10 @@ class TestACElectricityBus:
         ).set_index("Unnamed: 0")
         installed_capacity = float(energy_production_data["pv_plant_01"][INSTALLED_CAP])
         # adapt index
-        result_time_series_pv.index = input_time_series_pv_shortened.index
+        input_time_series_pv_shortened.index = result_time_series_pv.dropna().index
 
         assert_series_equal(
-            result_time_series_pv.astype(np.float64),
+            result_time_series_pv.astype(np.float64).dropna(),
             input_time_series_pv_shortened * installed_capacity,
             check_names=False,
         )
@@ -299,6 +299,7 @@ class TestACElectricityBus:
                 os.path.join(TEST_OUTPUT_PATH, case, "timeseries_all_busses.xlsx"),
                 sheet_name="Electricity",
             )
+            busses_flow.dropna(inplace=True)
             # compute the sum of the excess electricity for all timesteps
             excess[case] = sum(busses_flow["Electricity" + EXCESS_SINK])
         # compare the total excess electricity between the two cases